Moved sync classes to correct package.
