+++ platform-sdk/swirlds-platform-core/src/test/java/com/swirlds/platform/SyncManagerTest.java
@@ -36,6 +36,7 @@ import com.swirlds.common.system.events.BaseEvent;
 import com.swirlds.platform.components.CriticalQuorum;
 import com.swirlds.platform.components.EventCreationRules;
 import com.swirlds.platform.eventhandling.EventTransactionPool;
+import com.swirlds.platform.gossip.sync.SyncManagerImpl;
 import com.swirlds.platform.internal.EventImpl;
 import com.swirlds.platform.network.RandomGraph;
 import com.swirlds.platform.reconnect.FallenBehindManagerImpl;
